+++ CHANGELOG.md
@@ -85,6 +85,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /profile/:id (GET)
   - /profile/:id/unassigned_parameters
   - /profile/trimmed
+  - /regions/:id (GET, DELETE)
   - /regions/:region_name/phys_locations
   - /regions/name/:region_name
   - /riak/bucket/:bucket/key/:key/vault